📝 fix(protocol): "is an object", not "is not null", and pending_agents as u64
Two corrections to revision 7, both found by the projection rather than by
review of the document.

The client formula in the pending_agents section said
`answer_request != null` for "your turn". In a three-state field that is also
true when the field is absent, which means unchanged and includes "this daemon
never looked" — so a client following the document literally would announce
"your turn, here is the question" for a session it knows nothing about. The
formula is now a table over the three readings, and the paragraph that follows
says why the test is presence of the object and admits the first draft got it
wrong. Getting this wrong in the normative document is worse than getting it
wrong in code, because the code is a projection of it.

pending_agents was u32 while every other patched integer on this wire is u64,
six of them. No bug, just an inconsistency visible to anyone reading the struct.

+++ crates/violeet-proto/src/wire.rs
@@ -226,7 +226,7 @@ pub struct SessionUpdated {
     /// `state == "idle" && pending_agents == Some(Some(0))`, and a fabricated
     /// zero would answer yes for a session that is busy.
     #[serde(skip_serializing_if = "Option::is_none")]
-    pub pending_agents: Patch<u32>,
+    pub pending_agents: Patch<u64>,
 }
 
 /// The cap on `answer_request.question`, in bytes.
@@ -828,7 +828,7 @@ mod tests {
             "zero must not be omitted"
         );
         assert_eq!(
-            read_patch::<u32>(&j, "pending_agents"),
+            read_patch::<u64>(&j, "pending_agents"),
             Some(Some(0)),
             "zero must read back as zero, distinct from absent and from null"
         );
@@ -837,13 +837,13 @@ mod tests {
         let mut cleared = SessionUpdated::new("s1", now());
         cleared.pending_agents = Some(None);
         let j = json_of(&DaemonToApp::SessionUpdated(cleared));
-        assert_eq!(read_patch::<u32>(&j, "pending_agents"), Some(None));
+        assert_eq!(read_patch::<u64>(&j, "pending_agents"), Some(None));
 
         let j = json_of(&DaemonToApp::SessionUpdated(SessionUpdated::new(
             "s1",
             now(),
         )));
-        assert_eq!(read_patch::<u32>(&j, "pending_agents"), None);
+        assert_eq!(read_patch::<u64>(&j, "pending_agents"), None);
     }
 
     /// A patch that says only "asking" must still be worth sending.
